organize components by section
