+++ common/components/componentsBySection/FindProjects/ProjectSearchBar.jsx
